support scala-native
