+++ project/plugins.sbt
@@ -7,3 +7,5 @@ addSbtPlugin("org.xerial.sbt"     % "sbt-sonatype"    % "2.3")
 addSbtPlugin("pl.project13.scala" % "sbt-jmh"         % "0.3.4")
 addSbtPlugin("org.scoverage"      % "sbt-scoverage"   % "1.5.1")
 addSbtPlugin("org.portable-scala" % "sbt-scalajs-crossproject" % "0.4.0")
+addSbtPlugin("org.portable-scala" % "sbt-scala-native-crossproject" % "0.4.0")
+addSbtPlugin("org.scala-native"   % "sbt-scala-native" % "0.3.7")
